refactoring file organization according to three modules proposed
