+++ backend/product_management/app/go.mod
@@ -1,0 +1,1 @@
+module product_management/app
