+++ backend/product_management/app/models/Sneaker.go
