+++ backend/product_management/context/controllers/SneakerHandler.go
@@ -7,8 +7,8 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 
-	"antwood_team/product_management/src/database"
-	"antwood_team/product_management/src/models"
+	"product_management/app/database"
+	"product_management/app/models"
 )
 
 func InsertSneaker(c *fiber.Ctx) error {
