+++ backend/product_management/context/go.mod
@@ -1,0 +1,1 @@
+module product_management/context
