+++ backend/product_management/router/HomeRouter.go
@@ -1,4 +1,4 @@
-package routes
+package router
 
 import "github.com/gofiber/fiber/v2"
