+++ backend/product_management/router/go.mod
@@ -1,0 +1,1 @@
+module product_management/router
