+++ backend/product_management/test/go.mod
@@ -1,0 +1,1 @@
+module product_management/test
